*: initial commit
This is the initial commit of the flow3r firmware repository. You will
notice that it's at a later date than subsequent commits - that's
because it's the effect of a history rewrite.

The initial firmware repository started as an experimental Micropython
git fork. After a while we've decided that keeping the entirety of the
Micropython history in-repository was a bad idea, so we rewrote history.