usermodule/badge_link: fix UAF
We were keeping GCable pointers to JackPin inside static structures
without adding a GC root. That meant that the GC was free to collect
them and re-use memory for something else, leading to weird UAF bugs,
especially when soft resetting.

We could've added a GC root, but it's easier to just statically allocate
the JackPin objects, too.