ctx: update ctx.h
This update brings some fixes to excluding miniz from build and
enabling binary search for glyphs in the internal font engine.

Also make licensing clearer, ctx.h itself which is in this folder is
licensed under LGPLv3+, it is the micropython bindings - uctx which
live in ../micropython/usermodule/mp_uctx.c which are MIT licensed.