imu: read all sensors at once at 100 Hz
Introduces an update function to the BSP IMU implementation which
fetches the complete state of the BMI270 at once. This includes memory
range of the auxilary pressure sensor.

Introduces a task in st3m which updates the sensor readings every 10 ms.
Priority is chosen to be below captouch and audio.

BSP functions let the user know if there actually was a new reading on
the specific sensor. st3m always return the last measurement without
this extra information.