py: re-introduce os/app button naming
This finally bakes in some UX assumptions about the meaning of each
button.

The app button (by default the left button) is for exclusive use by
applications. In the menu, the app button is used to go left/right and
to select an option.

The os button (by default the right button) is for exclusive use by the
OS. It always adjusts the volume and allows the user to navigate back
(either exit the app or go back in menus - the latter not being
implemented yet).