st3m_leds: make LED indexing clockwise (api breaking)
Also:

- improved LED docs
- fixed the repr of the leds module
- fixed fix LED_POSITIONS in the sim

About the last point, the previous version was clockwise but not
starting from the top, but from the "first" LED according to the
hardware, which is ~11 LEDs after the one at the top. The C firmware
already adjusts for this.

Not fixed by this commit: Melodic demo (it's already a mess)

Didn't need changes: nick (now rotates clockwise)