bl00mbox: migration
bl00mbox and radspa have grown up. It is time for them to
gain independence in their very own repositories.

Check out gitlab.com/moon2embeddedaudio for future updates
or contributing.

We will continue to push updates to the flow3r project and
add hardware-specific features.

We see no reason to submodule it right away, but if it is
preferable for the flow3r firmware repo we're open to it.

Thanks for all the support and kind words, we'll try our
best to polish all the rough spots and add a ton of plugins
and features :D!