st3m: Mix media output together with bl00mbox
Not the cleanest thing, but it allows to use media framework
together with bl00mbox until a proper solution described in
https://git.flow3r.garden/flow3r/flow3r-firmware/-/issues/119#note_13069
is figured out.

Performance wise it appears to work pretty well - it doesn't
cause any troubles when playing video, and MOD & MP3 can even
comfortably play together with shoegaze.