py,ui: revert to old behavior of view transition blend
This works in most cases, but instead of the *whole* group first getting
composited and then adjusted it is the alpha of all individual draw calls
that are tweaked through the global alpha.

Issue #122, I am not sure compositing groups have ever worked properly - it is
a very little tested feature. The API was used wrong anyways the global opacity
changed the opacity of all render sources in the incoming view to have bits of
transparency, so there might have been no pixel compositing going on, or if it
was it was on pre-made-transparent bits instead of a texture being made
transparent during compositing.