uctx: deprecated start_group and end_group
They seem to not work as documented in the ctx protocol, in the cases where it
seemed to work it was due to side effects of global alpha, existing code should
continue working after this commit. The API might return, it is not removed
from the ctx protocol - and can still be accessed through ctx.parse().