st3m_gfx: change default graphics mode to 16bpp+osd+low_latency
This is a measure taken against issue #131 which seems to be a race/concurrent
access issue when OSD is combined with having 3 concurrent ctx drawlists in
flight.

For apps using the accelerometer or gyroscope to compute transforms/actions
this is an improvement, this might also generally be a benefit for many music
interfaces. In the cases where the deeper pipeline brings benefit apps can set
a custom graphics mode with deeper pipeline and no osd, though for instance in
the menus the frame rate achieved seems to now be the same for both types.