st3m_gfx: fix direct ctx variants
Using busy-polling to wait for blit to be done, using a mutex here
instead did not work as intended. The direct modes are good for debugging
and currently without a graphics task that floats between cores can
be used in scenarios where cpu load for audio is high enough to starve
graphics. The direct modes also are good in that you can immediately
read rendered pixels back from framebuffers after issuing drawing commands
and do not need to wait for the next frame.