py,st3m-ui: hold overlays mutex for shorter intervals
For fps overlay only show every 5th frame, for the regular overlay
first collect list of active overlays, and only if there is overlays
active get and flush the overlay ctx.

Also clearing a smaller rectangle with default overlays - the fps code
which does a lot less work is more reliable for use with mpremote.