st3m_gfx: Pin the graphics task to core 0
"graphics" task uses floating-point arithmetics, which makes it
pinned to a core by the scheduler. It appears that we were
lucky and it was ending up on core 0, but random changes can
make it land on core 1 together with micropython instead,
completely defeating the drawlist pipeline architecture
and making performance visibly suffer.

Explicitly pin the task to core 0, so we can be sure it lands
there no matter what.