py,st3m: Move WiFi and LED state handling out of Application class
Application class is the user's code, we shouldn't rely on it
implementing system-level functionality as there are several
ways for the user to accidentally break it.

This isn't a perfect place for it either, but it's at least better
than before. It's going to be revisited in the future, perhaps
by giving each app their own ViewManager instance that could
then properly track application's lifecycle.

One functional difference is that st3m.wifi._onoff_wifi_update()
is now being called when quitting any app, not just those
that request a particular WiFi state in their config.