python_payload: Drop super().on_{enter_done,exit} calls
They're not needed. The only reason to call super() on methods
in BaseView-descendants is to (optionally) use the features
provided by BaseView. These calls, while not causing harm, don't
provide any features to the implementer, so the implementer can't
rely on features that don't exist. Remove these calls to make
the code cleaner, less confusing and better serving as good
examples.