sensors: don't use temperature for altitude (also tiny math typo fix)
note: for accurate absolute altitude calculation we'd need to know the local
pressure at sea level, outside temperature and ideally relative humidity.
while it is certainly possible to retrieve them via wifi etc., it is not in
the scope of this application to do so.

relative altitude is proportional to outside temperature, so we introduce a
multiplicative error by ignoring it. this scale error is 7% at 35degC and
-4% at 0degC. in comparison, if we assume the temperature reading is the
outside temperature, going from a heated apartment to the balcony at same
elevation can cause a shift that may well be in the magnitude of 10m. we
find the former to be preferable to the latter.