py,st3m: Split SunMenu and MainMenu
SunMenu is just a presentation layer that doesn't concern itself with
any kind of logic. It could be simply replaced with (already existing)
SimpleMenu and FlowerMenu or, in the future, may even become customizable
by the user.

Split the main menu's logic into a separate entity that delegates
the presentation to the class it inherits from, like all the other
uses of the menu system.