captouch calibration routine: major bugfix
AFE would only get tweaked if it was necessary for the
last pad in the sequence. This can lead to random bad
calibration for any pads except for the last ones.

(also changed behavior to not autocalib in C on boot,
we're doing that OS-side anyways and alternate firmwares
also benefit from developers making an active choice.)