melodic demo: temporarily hide from menu
still didn't get around to finish it right, it's in an awkward
in-between state and does not really build on the differences
to scalar yet.

with a tear in our eye we send it off to a nap, but like a
worm emerges from its cocoon to grow into a butterfly this old
companion too will see the light of menu again more beautiful
than ever before. sweet dreams, it won't be long.